move extractSampleItemsDatums to _test.go
